pyproject.toml
[project]
name = "oraculus-db-masking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
